+++ __tests__/hooks/useEtherspotHistory.test.js
@@ -2,18 +2,14 @@ import { renderHook, waitFor } from '@testing-library/react';
 import { ethers } from 'ethers';
 
 // hooks
-import { useEtherspotHistory, EtherspotTransactionKit } from '../../src';
+import { EtherspotTransactionKit, useEtherspotHistory } from '../../src';
 
-const ethersProvider = new ethers.providers.JsonRpcProvider('http://localhost:8545', 'goerli'); // replace with your node's RPC URL
+const ethersProvider = new ethers.providers.JsonRpcProvider('http://localhost:8545', 'sepolia'); // replace with your node's RPC URL
 const provider = new ethers.Wallet.createRandom().connect(ethersProvider);
 
 describe('useEtherspotHistory()', () => {
   it('getAccountTransactions() returns account history', async () => {
-    const wrapper = ({ children }) => (
-      <EtherspotTransactionKit provider={provider}>
-        {children}
-      </EtherspotTransactionKit>
-    );
+    const wrapper = ({ children }) => <EtherspotTransactionKit provider={provider}>{children}</EtherspotTransactionKit>;
 
     const { result, rerender } = renderHook(({ chainId }) => useEtherspotHistory(chainId), {
       initialProps: { chainId: 1 },
@@ -43,11 +39,7 @@ describe('useEtherspotHistory()', () => {
     expect(accountTransactionsPolygon.length).toEqual(0);
   });
   it('getAccountTransaction() returns transaction by existing hash', async () => {
-    const wrapper = ({ children }) => (
-      <EtherspotTransactionKit provider={provider}>
-        {children}
-      </EtherspotTransactionKit>
-    );
+    const wrapper = ({ children }) => <EtherspotTransactionKit provider={provider}>{children}</EtherspotTransactionKit>;
 
     const { result, rerender } = renderHook(({ chainId }) => useEtherspotHistory(chainId), {
       initialProps: { chainId: 1 },
@@ -71,4 +63,52 @@ describe('useEtherspotHistory()', () => {
     const transactionPolygon = await result.current.getAccountTransaction('0x42');
     expect(transactionPolygon).toEqual(undefined);
   });
-})
+
+  it('returns transaction status for a given hash', async () => {
+    const transactionStatus = {
+      connextscanUrl: 'https://connextscan.io/tx/0x123',
+      status: 'completed',
+      transactionHash: '0x123',
+      transferId: 'abc123',
+    };
+
+    const wrapper = ({ children }) => <EtherspotTransactionKit provider={provider}>{children}</EtherspotTransactionKit>;
+
+    const { result } = renderHook(({ chainId }) => useEtherspotHistory(chainId), {
+      initialProps: { chainId: 1 },
+      wrapper,
+    });
+
+    // wait for history to be fetched for chain ID 1
+    await waitFor(() => expect(result.current).not.toBeNull());
+
+    // all correct props
+    const transaction1 = await result.current
+      .getAccountTransactionStatus(1, 137, '0x123')
+      .catch((e) => {
+        console.error(e);
+        return e;
+      });
+
+    expect(transaction1).toEqual(transactionStatus);
+
+    // transaction not found or does not exist
+    const transaction2 = await result.current
+      .getAccountTransactionStatus(4, 137, '0x123')
+      .catch((e) => {
+        console.error(e);
+        return e;
+      });
+
+    expect(transaction2).toEqual({})
+
+    // missing props
+    const transaction3 = await result.current.getAccountTransactionStatus(4, 137)
+      .catch((e) => {
+        console.error(e);
+        return e;
+      });
+
+    expect(transaction3).toBe('getTransactionStatus: missing required props')
+  });
+});