+++ __tests__/hooks/useEtherspotNfts.test.js
@@ -4,7 +4,7 @@ import { ethers } from 'ethers';
 // hooks
 import { useEtherspotNfts, EtherspotTransactionKit } from '../../src';
 
-const ethersProvider = new ethers.providers.JsonRpcProvider('http://localhost:8545', 'goerli'); // replace with your node's RPC URL
+const ethersProvider = new ethers.providers.JsonRpcProvider('http://localhost:8545', 'sepolia'); // replace with your node's RPC URL
 const provider = new ethers.Wallet.createRandom().connect(ethersProvider);
 
 describe('useEtherspotNfts()', () => {