+++ __tests__/hooks/useWalletAddress.test.js
@@ -9,7 +9,7 @@ import {
   otherFactoryDefaultAccountAddress,
 } from '../../__mocks__/@etherspot/prime-sdk';
 
-const ethersProvider = new ethers.providers.JsonRpcProvider('http://localhost:8545', 'goerli'); // replace with your node's RPC URL
+const ethersProvider = new ethers.providers.JsonRpcProvider('http://localhost:8545', 'sepolia'); // replace with your node's RPC URL
 const provider = new ethers.Wallet.createRandom().connect(ethersProvider);
 
 const providerWalletAddress = provider.address;